+++ Classes/Buildings/StorageBuilding.cpp
@@ -71,28 +71,28 @@ void StorageBuilding::setLevel(int level) {
 }
 
 float StorageBuilding::getCurrentMaxHP() const {
-    if (_level >= 0 && _level < _config->HP.size()) {
+    if (_level >= 0 && static_cast<size_t>(_level) < _config->HP.size()) {
         return _config->HP[_level];
     }
     return _config->HP.empty() ? 0.0f : _config->HP[0];
 }
 
 float StorageBuilding::getCurrentDP() const {
-    if (_level >= 0 && _level < _config->DP.size()) {
+    if (_level >= 0 && static_cast<size_t>(_level) < _config->DP.size()) {
         return _config->DP[_level];
     }
     return _config->DP.empty() ? 0.0f : _config->DP[0];
 }
 
 float StorageBuilding::getCurrentADD_STORAGE_ELIXIR_CAPACITY() const {
-    if (_level >= 0 && _level < _config->ADD_STORAGE_ELIXIR_CAPACITY.size()) {
+    if (_level >= 0 && static_cast<size_t>(_level) < _config->ADD_STORAGE_ELIXIR_CAPACITY.size()) {
         return _config->ADD_STORAGE_ELIXIR_CAPACITY[_level];
     }
     return _config->ADD_STORAGE_ELIXIR_CAPACITY.empty() ? 0.0f : _config->ADD_STORAGE_ELIXIR_CAPACITY[0];
 }
 
 float StorageBuilding::getCurrentADD_STORAGE_GOLD_CAPACITY() const {
-    if (_level >= 0 && _level < _config->ADD_STORAGE_GOLD_CAPACITY.size()) {
+    if (_level >= 0 && static_cast<size_t>(_level) < _config->ADD_STORAGE_GOLD_CAPACITY.size()) {
         return _config->ADD_STORAGE_GOLD_CAPACITY[_level];
     }
     return _config->ADD_STORAGE_GOLD_CAPACITY.empty() ? 0.0f : _config->ADD_STORAGE_GOLD_CAPACITY[0];
